Bugfix: reading temp folder from env for unzipping Added extra setting to use OS Temp folder instead of app folder

+++ ui/tabs/settings_tab.py
@@ -44,7 +44,9 @@ def settings_tab():
                 video_quality = gr.Slider(0, 100, value=roop.globals.CFG.video_quality, label="Video Quality (crf)", info='default: 14', step=1.0, interactive=True)
             with gr.Column():
                 button_apply_restart = gr.Button("Restart Server", variant='primary')
-                settings_controls.append(gr.Checkbox(label='Start with active live cam', value=roop.globals.CFG.live_cam_start_active, elem_id='live_cam_start_active', interactive=True))
+                with gr.Box():
+                    settings_controls.append(gr.Checkbox(label='Start with active live cam', value=roop.globals.CFG.live_cam_start_active, elem_id='live_cam_start_active', interactive=True))
+                    settings_controls.append(gr.Checkbox(label='Use OS temp folder', value=roop.globals.CFG.use_os_temp_folder, elem_id='use_os_temp_folder', interactive=True))
                 button_clean_temp = gr.Button("Clean temp folder")
                 button_apply_settings = gr.Button("Apply Settings")
 
@@ -100,7 +102,8 @@ def on_settings_changed(evt: gr.SelectData):
 def clean_temp():
     from ui.main import prepare_environment
     
-    shutil.rmtree(os.environ["TEMP"])
+    if not roop.globals.CFG.use_os_temp_folder:
+        shutil.rmtree(os.environ["TEMP"])
     prepare_environment()
    
     ui.globals.ui_input_thumbs.clear()
